Update file header with copyright notice
Fixes #2
Update the header of every Python source file with some boiler plate to indicate the authorship and licensing of the package.
In some cases, in files to do with PP for Quantum Espresso, this will be changed back to the standard AiiDA boiler plate as PP should ultimately be submitted as a contribution to aiida-quantumespresso via a pull request.
In all other cases, the URL will need updated when a GitHub page for the project has been registered.