Convert codebase to Python2/3 hybrid code
Converted the codebase to Python2/3 hybrid code in anticipation
of the end of Python2 support at the end of 2019. Hybrid code
that can run in Python2 and 3 is a necessary in AiiDA plugins
to ensure compatibility with any AiiDA setup. For example, if a
plugin used as a part of a work chain only supports Python 2, or only
Python 3, the whole setup will be forced to use that Python version.
Until AiiDA core drops Py2 support officially, all plugins should
maintain hybrid code.

The main changes in the fles are to use absolute imports and Py3 style
printing from future, and to use the six wrapper for dict, range, zip,
etc. The six imports can simply be removed when Py2 support is
dropped. For now, the six wrappers for dict are used, as these are by
default applied by the python-modernize tool. However, they are
ultimately not needed, and I will rewrite them in the process of
refactoring.