Fix bug for elongated cells
Fixes a bug where numpy crashes due to matricies being incompatible shapes.
When using meshgrid from numpy, by default, an ordering called 'xy' is used. This has the side effect of altering the ordering of dimensions in arrays leading to errors in linear algebra operations.
The solution is to specify 'ij' ordering, which keeps the indicies in the same order. This fix revealed an additional bug in the nested trapezium rule calculation of the energy. The axis over which to
do the trapezium rule should be specified explicitly to avoid broadcasting issues.