Reintroduce option to specify gaussian to use for model charge distribution
The GaussianCountercharge workchain had been limited to only using a fitted gaussian charge distribution.
In some cases, this fitting fails and the workchain cannot continue. It's desirable to be able to specify
a fixed width gaussian for stability. As the workchain already can generate a model gaussian of any shape
via 9 parameters (which are found by fitting), now the user can specify these directly.